ml-service/requirements.txt
flask==2.3.2
tensorflow>=2.20.0
numpy>=1.24.3
pillow>=10.0.0
requests>=2.31.0
inference-sdk>=0.62.2